Get bert benchmark to work again
The code to re-wire inplace ops was assuming that the first input had the same size as the output.
After debugging the failure in BERT, I found that the inplace op that threw the assertion error was a add operation where one input had a smaller size than the other. (That's probably a broadcast add).
So I modified the condition to look for the input that has the same size as the output, and re-wire that instead.